Extract classes from course file
